Factor out general purpose functions
