+++ hibike/unit_tests/hibike_message_tests.py
@@ -13,19 +13,8 @@ import serial
 
 import hibike_message
 import spawn_virtual_devices
+from unit_tests.utils import run_with_random_data
 DEVICE_TYPES = list(hibike_message.DEVICES)
-
-
-def run_with_random_data(func, arg_func, kwarg_func=lambda: {}, times=5):
-    """
-    Test FUNC with random arguments generated by ARG_FUNC and
-    KWARG_FUNC, repeating TIMES.
-
-    ARG_FUNC should return an iterable.
-    """
-    for _ in range(times):
-        # pylint: disable=syntax-error
-        func(*arg_func(), **kwarg_func())
 
 
 def random_uid():
@@ -44,6 +33,7 @@ def random_params(device_id):
     sample_length = random.randrange(len(all_params) + 1)
     params = random.sample(all_params, sample_length)
     return params
+
 
 def random_values(device_id, params):
     """
